Initial commit!
This version is able to produce a nix derivation which is perilously
close to being a valid quicklisp dist.  The dist only contains the
systems and releases which are required.  Then, we'll use quicklisp's
bundle operation to create a bundle that contains all the systems
along with a convenient loader file.  We'll stash that bundle into the
nix store and create a script that runs the lisp compiler, loads the
loader file, and then... well... does whatever the user wants.  At
that point, the user will have all of the desired dependencies
available to them!