Remove nix-safe-name!
The attribute sets we're generating don't need to be convenient to use
with tools like nix-env and nix-shell.  Let's just quote them and be
done with it!

We should still be escaping nix control characters in the system names
and such, though.  Imagine if a system was named
    evilKey${evilLogic}" = moreEvil; "realName
So, yeah.  We're currently vulnerable to injection attacks.